SepBy -> Ski
Separating Outer from Inner did not bring the benfits I expected; in
Symmetric.Free, I do have a similar distinction, but they type of the
index between Inner and Outer is different from the type of the index
between Outer and Intter (which in turn is the same as the type of the
index between Outer and Final). I could make SepBy more complex by
adding more index types, but I think it would be simpler to encapsulate
the Outer+Inner piece into a single Q which callers may or may not
choose to instantiate to a type consisting of an Outer followed by an
Inner.