StringDiagram
like Ski, StringDiagram is a generic implementation which is intended to
simplify later proofs, by providing a generic type with generic proofs,
which can later be specialized via smaller types and smaller proofs.

StringDiagram is a lot more specialized than Ski: there are lines, the
lines can cross, at each step we focus on one set of lines and transform
them into some other set of lines, and the leftover lines are left
alone, to be used during a later step. What varies is how we focus and
rearrange the lines.