simplify by solving more general problems
compose-Ski's key insight is to propagate the Final value through the
second Ski. I extracted this idea to propagate-Ski in the hope that it
will be used elsewhere, and that allowed me to simplify compose-Ski a
little bit.

widen-StringDiagram was big and complicated, it didn't really have a
core idea. I generalized it to bump-StringDiagram, in which we perform
an arbitrary transformation to the list, not just widening, and now the
complexity is more evenly split between bump-StringDiagram and
widen-StringDiagram.